requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=1.0.0
openai>=1.40.0
pydantic>=2.7.4
pydantic-settings>=2.10.1
python-dotenv==1.0.0
httpx==0.25.2
pillow==10.1.0
python-multipart==0.0.6
python-json-logger==2.0.7
